+++ src/types/couponInterface.ts
@@ -1,0 +1,9 @@
+export interface CouponAttributes {
+  id: number;
+  code: string;
+  discount: number;
+  isActive: boolean;
+  expiredDate: Date;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
